Updates to the Demos so they are mains. Moving the reader, writer and round tripper to utils clas
